-Split Int256 file. -Fixed BigInteger truncating and saturating conversions.
