+++ src/MissingValues/Int256.GenericMath.cs
@@ -1,15 +1,9 @@
 ﻿using MissingValues.Info;
 using MissingValues.Internals;
-using System.Buffers.Binary;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Numerics;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Runtime.Intrinsics;
-using System.Runtime.Intrinsics.X86;
-using System.Text;
 using MissingValues.Primitives;
 
 namespace MissingValues
@@ -21,12 +15,6 @@ namespace MissingValues
 		IPowerFunctions<Int256>,
 		IFormattableSignedInteger<Int256>
 	{
-		private static UInt128 _upperMin => new UInt128(0x8000_0000_0000_0000, 0x0000_0000_0000_0000);
-		private static UInt128 _lowerMin => new UInt128(0x0000_0000_0000_0000, 0x0000_0000_0000_0000);
-
-		private static UInt128 _upperMax => new UInt128(0x7FFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF);
-		private static UInt128 _lowerMax => new UInt128(0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF);
-
 		static Int256 IBinaryNumber<Int256>.AllBitsSet => new(_lowerMax, _lowerMax);
 		// MaxValue = 57896044618658097711785492504343953926634992332820282019728792003956564819967
 		static Int256 IMinMaxValue<Int256>.MaxValue => MaxValue;
@@ -648,220 +636,6 @@ namespace MissingValues
 			return true;
 		}
 
-		static bool INumberBase<Int256>.TryConvertFromChecked<TOther>(TOther value, out Int256 result) => TryConvertFromChecked(value, out result);
-		private static bool TryConvertFromChecked<TOther>(TOther value, out Int256 result)
-			where TOther : INumberBase<TOther>
-		{
-			bool converted = true;
-
-			checked
-			{
-				result = value switch
-				{
-					char actual => (Int256)actual,
-					Half actual => (Int256)actual,
-					float actual => (Int256)actual,
-					double actual => (Int256)actual,
-					decimal actual => (Int256)actual,
-					byte actual => (Int256)actual,
-					ushort actual => (Int256)actual,
-					uint actual => (Int256)actual,
-					ulong actual => (Int256)actual,
-					UInt128 actual => (Int256)actual,
-					UInt256 actual => (Int256)actual,
-					UInt512 actual => (Int256)actual,
-					nuint actual => (Int256)actual,
-					sbyte actual => (Int256)actual,
-					short actual => (Int256)actual,
-					int actual => (Int256)actual,
-					long actual => (Int256)actual,
-					Int128 actual => (Int256)actual,
-					Int256 actual => actual,
-					Int512 actual => (Int256)actual,
-					nint actual => (Int256)actual,
-					BigInteger actual => (Int256)actual,
-					_ => BitHelper.DefaultConvert<Int256>(out converted)
-				};
-			}
-
-			return converted;
-		}
-
-		static bool INumberBase<Int256>.TryConvertFromSaturating<TOther>(TOther value, out Int256 result) => TryConvertFromSaturating(value, out result);
-		private static bool TryConvertFromSaturating<TOther>(TOther value, out Int256 result)
-			where TOther : INumberBase<TOther>
-		{
-			const double TwoPow255 = 57896044618658097711785492504343953926634992332820282019728792003956564819968.0;
-
-			bool converted = true;
-			result = value switch
-			{
-				char actual => actual,
-				Half actual => (Int256)actual,
-				float actual => (Int256)actual,
-				double actual => (actual <= -TwoPow255) ? MinValue : (actual > +TwoPow255) ? MaxValue : (Int256)actual,
-				decimal actual => (Int256)actual,
-				byte actual => (Int256)actual,
-				ushort actual => (Int256)actual,
-				uint actual => (Int256)actual,
-				ulong actual => (Int256)actual,
-				UInt128 actual => (Int256)actual,
-				UInt256 actual => (actual > (UInt256)MaxValue) ? MaxValue : (Int256)actual,
-				nuint actual => (Int256)actual,
-				sbyte actual => actual,
-				short actual => actual,
-				int actual => actual,
-				long actual => actual,
-				Int128 actual => actual,
-				Int256 actual => actual,
-				Int512 actual => (actual < MinValue) ? MinValue : (actual > MaxValue) ? MaxValue : (Int256)actual,
-				nint actual => actual,
-				BigInteger actual => (actual < (BigInteger)MinValue) ? MinValue : (actual > (BigInteger)MaxValue) ? MaxValue : (Int256)actual,
-				_ => BitHelper.DefaultConvert<Int256>(out converted)
-			};
-			return converted;
-		}
-
-		static bool INumberBase<Int256>.TryConvertFromTruncating<TOther>(TOther value, out Int256 result) => TryConvertFromTruncating(value, out result);
-		private static bool TryConvertFromTruncating<TOther>(TOther value, out Int256 result)
-			where TOther : INumberBase<TOther>
-		{
-			const double TwoPow255 = 57896044618658097711785492504343953926634992332820282019728792003956564819968.0;
-
-			bool converted = true;
-			result = value switch
-			{
-				char actual => actual,
-				Half actual => (Half.IsPositiveInfinity(actual)) ? MaxValue : (Half.IsNegativeInfinity(actual)) ? MinValue : (Int256)actual,
-				float actual => (float.IsPositiveInfinity(actual)) ? MaxValue : (float.IsNegativeInfinity(actual)) ? MinValue : (Int256)actual,
-				double actual => (actual <= -TwoPow255) ? MinValue : (actual > +TwoPow255) ? MaxValue : (Int256)actual,
-				decimal actual => (Int128)actual,
-				byte actual => (Int256)actual,
-				ushort actual => (Int256)actual,
-				uint actual => (Int256)actual,
-				ulong actual => (Int256)actual,
-				UInt128 actual => (Int256)actual,
-				UInt256 actual => (Int256)actual,
-				nuint actual => (Int256)actual,
-				sbyte actual => actual,
-				short actual => actual,
-				int actual => actual,
-				long actual => actual,
-				Int128 actual => actual,
-				Int256 actual => actual,
-				nint actual => actual,
-				BigInteger actual => (Int256)actual,
-				_ => BitHelper.DefaultConvert<Int256>(out converted)
-			};
-			return converted;
-		}
-
-		static bool INumberBase<Int256>.TryConvertToChecked<TOther>(Int256 value, out TOther result)
-		{
-			bool converted = true;
-			result = TOther.Zero;
-			checked
-			{
-				result = result switch
-				{
-					char => (TOther)(object)(char)value,
-					Half => (TOther)(object)(Half)value,
-					float => (TOther)(object)(float)value,
-					double => (TOther)(object)(double)value,
-					decimal => (TOther)(object)(decimal)value,
-					byte => (TOther)(object)(byte)value,
-					ushort => (TOther)(object)(ushort)value,
-					uint => (TOther)(object)(uint)value,
-					ulong => (TOther)(object)(ulong)value,
-					UInt128 => (TOther)(object)(UInt128)value,
-					UInt256 => (TOther)(object)(UInt256)value,
-					UInt512 => (TOther)(object)(UInt512)value,
-					nuint => (TOther)(object)(nuint)value,
-					sbyte => (TOther)(object)(sbyte)value,
-					short => (TOther)(object)(short)value,
-					int => (TOther)(object)(int)value,
-					long => (TOther)(object)(long)value,
-					Int128 => (TOther)(object)(Int128)value,
-					Int256 => (TOther)(object)value,
-					Int512 => (TOther)(object)(Int512)value,
-					nint => (TOther)(object)(nint)value,
-					BigInteger => (TOther)(object)(BigInteger)value,
-					_ => BitHelper.DefaultConvert<TOther>(out converted)
-				};
-			}
-
-			return converted;
-		}
-
-		static bool INumberBase<Int256>.TryConvertToSaturating<TOther>(Int256 value, out TOther result)
-		{
-			bool converted = true;
-			result = TOther.Zero;
-
-			result = result switch
-			{
-				char => (TOther)(object)(char)value,
-				Half => (TOther)(object)(Half)value,
-				float => (TOther)(object)(float)value,
-				double => (TOther)(object)(double)value,
-				decimal => (TOther)(object)(decimal)value,
-				byte => (TOther)(object)((value >= (Int256)byte.MaxValue) ? byte.MaxValue : (value <= (Int256)byte.MinValue) ? byte.MinValue : (byte)value),
-				ushort => (TOther)(object)((value >= (Int256)ushort.MaxValue) ? ushort.MaxValue : (value <= (Int256)ushort.MinValue) ? ushort.MinValue : (ushort)value),
-				uint => (TOther)(object)((value >= (Int256)uint.MaxValue) ? uint.MaxValue : (value <= (Int256)uint.MinValue) ? uint.MinValue : (uint)value),
-				ulong => (TOther)(object)((value >= (Int256)ulong.MaxValue) ? ulong.MaxValue : (value <= (Int256)ulong.MinValue) ? ulong.MinValue : (ulong)value),
-				UInt128 => (TOther)(object)((value >= (Int256)UInt128.MaxValue) ? UInt128.MaxValue : (value <= (Int256)UInt128.MinValue) ? UInt128.MinValue : (UInt128)value),
-				UInt256 => (TOther)(object)(UInt256)value,
-				UInt512 => (TOther)(object)(UInt512)value,
-				nuint => (TOther)(object)((value >= (Int256)nuint.MaxValue) ? nuint.MaxValue : (value <= (Int256)nuint.MinValue) ? nuint.MinValue : (nuint)value),
-				sbyte => (TOther)(object)((value >= (Int256)sbyte.MaxValue) ? sbyte.MaxValue : (value <= (Int256)sbyte.MinValue) ? sbyte.MinValue : (sbyte)value),
-				short => (TOther)(object)((value >= (Int256)short.MaxValue) ? short.MaxValue : (value <= (Int256)short.MinValue) ? short.MinValue : (short)value),
-				int => (TOther)(object)((value >= (Int256)int.MaxValue) ? int.MaxValue : (value <= (Int256)int.MinValue) ? int.MinValue : (int)value),
-				long => (TOther)(object)((value >= (Int256)long.MaxValue) ? long.MaxValue : (value <= (Int256)long.MinValue) ? long.MinValue : (long)value),
-				Int128 => (TOther)(object)((value >= (Int256)Int128.MaxValue) ? Int128.MaxValue : (value <= (Int256)Int128.MinValue) ? Int128.MinValue : (Int128)value),
-				Int256 => (TOther)(object)value,
-				Int512 => (TOther)(object)(Int512)value,
-				nint => (TOther)(object)((value >= (Int256)nint.MaxValue) ? nint.MaxValue : (value <= (Int256)nint.MinValue) ? nint.MinValue : (nint)value),
-				BigInteger => (TOther)(object)(BigInteger)value,
-				_ => BitHelper.DefaultConvert<TOther>(out converted)
-			};
-
-			return converted;
-		}
-
-		static bool INumberBase<Int256>.TryConvertToTruncating<TOther>(Int256 value, out TOther result)
-		{
-			bool converted = true;
-			result = TOther.Zero;
-			result = result switch
-			{
-				char => (TOther)(object)(char)value,
-				Half => (TOther)(object)(Half)value,
-				float => (TOther)(object)(float)value,
-				double => (TOther)(object)(double)value,
-				decimal => (TOther)(object)(decimal)value,
-				byte => (TOther)(object)(byte)value,
-				ushort => (TOther)(object)(ushort)value,
-				uint => (TOther)(object)(uint)value,
-				ulong => (TOther)(object)(ulong)value,
-				UInt128 => (TOther)(object)(UInt128)value,
-				UInt256 => (TOther)(object)(UInt256)value,
-				UInt512 => (TOther)(object)(UInt512)value,
-				nuint => (TOther)(object)(nuint)value,
-				sbyte => (TOther)(object)(sbyte)value,
-				short => (TOther)(object)(short)value,
-				int => (TOther)(object)(int)value,
-				long => (TOther)(object)(long)value,
-				Int128 => (TOther)(object)(Int128)value,
-				Int256 => (TOther)(object)value,
-				Int512 => (TOther)(object)(Int512)value,
-				nint => (TOther)(object)(nint)value,
-				BigInteger => (TOther)(object)(BigInteger)value,
-				_ => BitHelper.DefaultConvert<TOther>(out converted)
-			};
-
-			return converted;
-		}
-
 		/// <inheritdoc/>
 		public int CompareTo(object? obj)
 		{
@@ -898,60 +672,6 @@ namespace MissingValues
 		}
 		
 		static Int256 IBigInteger<Int256>.Create(ReadOnlySpan<ulong> parts) => new(parts);
-
-		/// <inheritdoc/>
-		public static Int256 CreateChecked<TOther>(TOther value)
-			where TOther : INumberBase<TOther>
-		{
-			Int256 result;
-
-			if (value is Int256 v)
-			{
-				result = v;
-			}
-			else if (!Int256.TryConvertFromChecked(value, out result) && !TOther.TryConvertToChecked<Int256>(value, out result))
-			{
-				Thrower.NotSupported<Int256, TOther>();
-			}
-
-			return result;
-		}
-
-		/// <inheritdoc/>
-		public static Int256 CreateSaturating<TOther>(TOther value)
-			where TOther : INumberBase<TOther>
-		{
-			Int256 result;
-
-			if (value is Int256 v)
-			{
-				result = v;
-			}
-			else if (!Int256.TryConvertFromSaturating(value, out result) && !TOther.TryConvertToSaturating<Int256>(value, out result))
-			{
-				Thrower.NotSupported<Int256, TOther>();
-			}
-
-			return result;
-		}
-
-		/// <inheritdoc/>
-		public static Int256 CreateTruncating<TOther>(TOther value)
-			where TOther : INumberBase<TOther>
-		{
-			Int256 result;
-
-			if (value is Int256 v)
-			{
-				result = v;
-			}
-			else if (!TryConvertFromTruncating(value, out result) && !TOther.TryConvertToTruncating(value, out result))
-			{
-				Thrower.NotSupported<Int256, TOther>();
-			}
-
-			return result;
-		}
 
 		/// <inheritdoc/>
 		public bool Equals(Int256 other)
@@ -1043,504 +763,5 @@ namespace MissingValues
 		static int IFormattableInteger<Int256>.UnsignedCompare(in Int256 value1, in Int256 value2) => unchecked(((UInt256)value1).CompareTo((UInt256)value2));
 		static int IFormattableInteger<Int256>.Log2Int32(in Int256 value) => BitHelper.Log2(in value);
 		static int IFormattableInteger<Int256>.LeadingZeroCountInt32(in Int256 value) => BitHelper.LeadingZeroCount(in value);
-
-		/// <inheritdoc/>
-		public static Int256 operator +(in Int256 value) => value;
-
-		/// <inheritdoc/>
-		public static Int256 operator +(in Int256 left, in Int256 right)
-		{
-			// For unsigned addition, we can detect overflow by checking `(x + y) < x`
-			// This gives us the carry to add to upper to compute the correct result
-
-			ulong part0 = left._p0 + right._p0;
-			ulong carry = (part0 < left._p0) ? 1UL : 0UL;
-
-			ulong part1 = left._p1 + right._p1 + carry;
-			carry = (part1 < left._p1 || (carry == 1 && part1 == left._p1)) ? 1UL : 0UL;
-
-			ulong part2 = left._p2 + right._p2 + carry;
-			carry = (part2 < left._p2 || (carry == 1 && part2 == left._p2)) ? 1UL : 0UL;
-
-			ulong part3 = left._p3 + right._p3 + carry;
-
-			return new Int256(part3, part2, part1, part0);
-		}
-		/// <inheritdoc/>
-		public static Int256 operator checked +(in Int256 left, in Int256 right)
-		{
-			// For signed addition, we can detect overflow by checking if the sign of
-			// both inputs are the same and then if that differs from the sign of the
-			// output.
-
-			Int256 result = left + right;
-
-			if ((long)((result._p3 ^ left._p3) & ~(left._p3 ^ right._p3)) < 0)
-			{
-				Thrower.ArithmeticOverflow(Thrower.ArithmeticOperation.Addition);
-			}
-			return result;
-		}
-
-		/// <inheritdoc/>
-		public static Int256 operator -(in Int256 value) => Zero - value;
-		/// <inheritdoc/>
-		public static Int256 operator checked -(in Int256 value) => checked(Zero - value);
-
-		/// <inheritdoc/>
-		public static Int256 operator -(in Int256 left, in Int256 right)
-		{
-			// For unsigned subtract, we can detect overflow by checking `(x - y) > x`
-			// This gives us the borrow to subtract from upper to compute the correct result
-
-			ulong part0 = left._p0 - right._p0;
-			ulong borrow = (part0 > left._p0) ? 1UL : 0UL;
-
-			ulong part1 = left._p1 - right._p1 - borrow;
-			borrow = (part1 > left._p1 || (borrow == 1UL && part1 == left._p1)) ? 1UL : 0UL;
-
-			ulong part2 = left._p2 - right._p2 - borrow;
-			borrow = (part2 > left._p2 || (borrow == 1UL && part2 == left._p2)) ? 1UL : 0UL;
-
-			ulong part3 = left._p3 - right._p3 - borrow;
-
-			return new Int256(part3, part2, part1, part0);
-		}
-		/// <inheritdoc/>
-		public static Int256 operator checked -(in Int256 left, in Int256 right)
-		{
-			// For signed subtraction, we can detect overflow by checking if the sign of
-			// both inputs are different and then if that differs from the sign of the
-			// output.
-
-			Int256 result = left - right;
-
-			uint sign = (uint)(left._p3 >> 63);
-
-			if (sign != (uint)(right._p3 >> 63) && sign != (uint)(result._p3 >> 63))
-			{
-				Thrower.ArithmeticOverflow(Thrower.ArithmeticOperation.Subtraction);
-			}
-			return result;
-		}
-
-		/// <inheritdoc/>
-		public static Int256 operator ~(in Int256 value)
-		{
-			if (Vector256.IsHardwareAccelerated)
-			{
-				var v = Unsafe.BitCast<Int256, Vector256<ulong>>(value);
-				var result = ~v;
-				return Unsafe.BitCast<Vector256<ulong>, Int256>(result);
-			}
-			else
-			{
-				return new(~value._p3, ~value._p2, ~value._p1, ~value._p0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static Int256 operator ++(in Int256 value) => value + One;
-		/// <inheritdoc/>
-		public static Int256 operator checked ++(in Int256 value) => checked(value + One);
-
-		/// <inheritdoc/>
-		public static Int256 operator --(in Int256 value) => value - One;
-		/// <inheritdoc/>
-		public static Int256 operator checked --(in Int256 value) => checked(value - One);
-
-		/// <inheritdoc/>
-		public static Int256 operator *(in Int256 left, in Int256 right) => (Int256)((UInt256)(left) * (UInt256)(right));
-		/// <inheritdoc/>
-		public static Int256 operator checked *(in Int256 left, in Int256 right)
-		{
-			Int256 upper = BigMul(left, right, out Int256 lower);
-
-			if (((upper != Zero) || (lower < Zero)) && ((~upper != Zero) || (lower >= Zero)))
-			{
-				// The upper bits can safely be either Zero or AllBitsSet
-				// where the former represents a positive value and the
-				// latter a negative value.
-				//
-				// However, when the upper bits are Zero, we also need to
-				// confirm the lower bits are positive, otherwise we have
-				// a positive value greater than MaxValue and should throw
-				//
-				// Likewise, when the upper bits are AllBitsSet, we also
-				// need to confirm the lower bits are negative, otherwise
-				// we have a large negative value less than MinValue and
-				// should throw.
-
-				Thrower.ArithmeticOverflow(Thrower.ArithmeticOperation.Multiplication);
-			}
-
-			return lower;
-		}
-
-		/// <inheritdoc/>
-		public static Int256 operator /(in Int256 left, in Int256 right)
-		{
-			DivRem(in left, in right, out Int256 quotient, out _);
-			return quotient;
-		}
-
-		/// <inheritdoc/>
-		public static Int256 operator checked /(in Int256 left, in Int256 right) => left / right;
-
-		/// <inheritdoc/>
-		public static Int256 operator %(in Int256 left, in Int256 right)
-		{
-			DivRem(in left, in right, out _, out Int256 remainder);
-			return remainder;
-		}
-
-		/// <inheritdoc/>
-		public static Int256 operator &(in Int256 left, in Int256 right)
-		{
-			if (Vector256.IsHardwareAccelerated)
-			{
-				var v1 = Unsafe.BitCast<Int256, Vector256<ulong>>(left);
-				var v2 = Unsafe.BitCast<Int256, Vector256<ulong>>(right);
-				var result = v1 & v2;
-				return Unsafe.BitCast<Vector256<ulong>, Int256>(result);
-			}
-			else if (Avx2.IsSupported)
-			{
-				var v1 = Unsafe.BitCast<Int256, Vector256<ulong>>(left);
-				var v2 = Unsafe.BitCast<Int256, Vector256<ulong>>(right);
-				var result = Avx2.And(v1, v2);
-				return Unsafe.BitCast<Vector256<ulong>, Int256>(result);
-			}
-			else
-			{
-				return new(left._p3 & right._p3, left._p2 & right._p2, left._p1 & right._p1, left._p0 & right._p0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static Int256 operator |(in Int256 left, in Int256 right)
-		{
-			if (Vector256.IsHardwareAccelerated)
-			{
-				var v1 = Unsafe.BitCast<Int256, Vector256<ulong>>(left);
-				var v2 = Unsafe.BitCast<Int256, Vector256<ulong>>(right);
-				var result = v1 | v2;
-				return Unsafe.BitCast<Vector256<ulong>, Int256>(result);
-			}
-			else if (Avx2.IsSupported)
-			{
-				var v1 = Unsafe.BitCast<Int256, Vector256<ulong>>(left);
-				var v2 = Unsafe.BitCast<Int256, Vector256<ulong>>(right);
-				var result = Avx2.Or(v1, v2);
-				return Unsafe.BitCast<Vector256<ulong>, Int256>(result);
-			}
-			else
-			{
-				return new(left._p3 | right._p3, left._p2 | right._p2, left._p1 | right._p1, left._p0 | right._p0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static Int256 operator ^(in Int256 left, in Int256 right)
-		{
-			if (Vector256.IsHardwareAccelerated)
-			{
-				var v1 = Unsafe.BitCast<Int256, Vector256<ulong>>(left);
-				var v2 = Unsafe.BitCast<Int256, Vector256<ulong>>(right);
-				var result = v1 ^ v2;
-				return Unsafe.BitCast<Vector256<ulong>, Int256>(result);
-			}
-			else if (Avx2.IsSupported)
-			{
-				var v1 = Unsafe.BitCast<Int256, Vector256<ulong>>(left);
-				var v2 = Unsafe.BitCast<Int256, Vector256<ulong>>(right);
-				var result = Avx2.Xor(v1, v2);
-				return Unsafe.BitCast<Vector256<ulong>, Int256>(result);
-			}
-			else
-			{
-				return new(left._p3 ^ right._p3, left._p2 ^ right._p2, left._p1 ^ right._p1, left._p0 ^ right._p0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static Int256 operator <<(in Int256 value, int shiftAmount)
-		{
-			// C# automatically masks the shift amount for UInt64 to be 0x3F. So we
-			// need to specially handle things if the shift amount exceeds 0x3F.
-
-			shiftAmount &= 0xFF; // mask the shift amount to be within [0, 255]
-
-			if (shiftAmount == 0)
-			{
-				return value;
-			}
-
-			if (shiftAmount < 64)
-			{
-				ulong part3 = (value._p3 << shiftAmount) | (value._p2 >> (64 - shiftAmount));
-				ulong part2 = (value._p2 << shiftAmount) | (value._p1 >> (64 - shiftAmount));
-				ulong part1 = (value._p1 << shiftAmount) | (value._p0 >> (64 - shiftAmount));
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new Int256(part3, part2, part1, part0);
-			}
-			else if (shiftAmount < 128)
-			{
-				shiftAmount -= 64;
-
-				if (shiftAmount == 0)
-				{
-					return new Int256(value._p2, value._p1, value._p0, 0);
-				}
-
-				ulong part2 = (value._p2 << shiftAmount) | (value._p1 >> (64 - shiftAmount));
-				ulong part1 = (value._p1 << shiftAmount) | (value._p0 >> (64 - shiftAmount));
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new Int256(part2, part1, part0, 0);
-			}
-			else if (shiftAmount < 192)
-			{
-				shiftAmount -= 128;
-
-				if (shiftAmount == 0)
-				{
-					return new Int256(value._p1, value._p0, 0, 0);
-				}
-
-				ulong part1 = (value._p1 << shiftAmount) | (value._p0 >> (64 - shiftAmount));
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new Int256(part1, part0, 0, 0);
-			}
-			else // shiftAmount < 256
-			{
-				shiftAmount -= 192;
-
-				if (shiftAmount == 0)
-				{
-					return new Int256(value._p0, 0, 0, 0);
-				}
-
-				ulong part0 = value._p0 << shiftAmount;
-
-				return new Int256(part0, 0, 0, 0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static Int256 operator >>(in Int256 value, int shiftAmount)
-		{
-			// need to specially handle things if the 15th bit is set.
-
-			shiftAmount &= 0xFF;
-
-			if (shiftAmount == 0)
-			{
-				return value;
-			}
-
-			if (shiftAmount < 64)
-			{
-				ulong part0 = (value._p0 >> shiftAmount) | (value._p1 << (64 - shiftAmount));
-				ulong part1 = (value._p1 >> shiftAmount) | (value._p2 << (64 - shiftAmount));
-				ulong part2 = (value._p2 >> shiftAmount) | (value._p3 << (64 - shiftAmount));
-				ulong part3 = (ulong)((long)value._p3 >> shiftAmount);
-
-				return new Int256(part3, part2, part1, part0);
-			}
-
-			ulong preservedSign = (ulong)((long)value._p3 >> 63);
-
-			if (shiftAmount < 128)
-			{
-				shiftAmount -= 64;
-
-				if (shiftAmount == 0)
-				{
-					return new Int256(preservedSign, value._p3, value._p2, value._p1);
-				}
-
-				ulong part0 = (value._p1 >> shiftAmount) | (value._p2 << (64 - shiftAmount));
-				ulong part1 = (value._p2 >> shiftAmount) | (value._p3 << (64 - shiftAmount));
-				ulong part2 = (ulong)((long)value._p3 >> shiftAmount);
-
-				return new Int256(preservedSign, part2, part1, part0);
-			}
-			else if (shiftAmount < 192)
-			{
-				shiftAmount -= 128;
-
-				if (shiftAmount == 0)
-				{
-					return new Int256(preservedSign, preservedSign, value._p3, value._p2);
-				}
-
-				ulong part0 = (value._p2 >> shiftAmount) | (value._p3 << (64 - shiftAmount));
-				ulong part1 = (ulong)((long)value._p3 >> shiftAmount);
-
-				return new Int256(preservedSign, preservedSign, part1, part0);
-			}
-			else // shiftAmount < 256
-			{
-				shiftAmount -= 192;
-
-				ulong part0 = (ulong)((long)value._p3 >> shiftAmount);
-
-				return new Int256(preservedSign, preservedSign, preservedSign, part0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static Int256 operator >>>(in Int256 value, int shiftAmount)
-		{
-			// C# automatically masks the shift amount for UInt64 to be 0x3F. So we
-			// need to specially handle things if the shift amount exceeds 0x3F.
-
-			shiftAmount &= 0xFF; // mask the shift amount to be within [0, 255]
-
-			if (shiftAmount == 0)
-			{
-				return value;
-			}
-
-			if (shiftAmount < 64)
-			{
-				ulong part0 = (value._p0 >> shiftAmount) | (value._p1 << (64 - shiftAmount));
-				ulong part1 = (value._p1 >> shiftAmount) | (value._p2 << (64 - shiftAmount));
-				ulong part2 = (value._p2 >> shiftAmount) | (value._p3 << (64 - shiftAmount));
-				ulong part3 = value._p3 >> shiftAmount;
-
-				return new Int256(part3, part2, part1, part0);
-			}
-			else if (shiftAmount < 128)
-			{
-				shiftAmount -= 64;
-
-				if (shiftAmount == 0)
-				{
-					return new Int256(0, value._p3, value._p2, value._p1);
-				}
-
-				ulong part0 = (value._p1 >> shiftAmount) | (value._p2 << (64 - shiftAmount));
-				ulong part1 = (value._p2 >> shiftAmount) | (value._p3 << (64 - shiftAmount));
-				ulong part2 = value._p3 >> shiftAmount;
-
-				return new Int256(0, part2, part1, part0);
-			}
-			else if (shiftAmount < 192)
-			{
-				shiftAmount -= 128;
-
-				if (shiftAmount == 0)
-				{
-					return new Int256(0, 0, value._p3, value._p2);
-				}
-
-				ulong part0 = (value._p2 >> shiftAmount) | (value._p3 << (64 - shiftAmount));
-				ulong part1 = value._p3 >> shiftAmount;
-
-				return new Int256(0, 0, part1, part0);
-			}
-			else // shiftAmount < 256
-			{
-				shiftAmount -= 192;
-
-				ulong part0 = value._p3 >> shiftAmount;
-
-				return new Int256(0, 0, 0, part0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static bool operator ==(in Int256 left, in Int256 right)
-		{
-			if (Vector256.IsHardwareAccelerated)
-			{
-				var v1 = Vector256.Create(left._p0, left._p1, left._p2, left._p3);
-				var v2 = Vector256.Create(right._p0, right._p1, right._p2, right._p3);
-				return v1 == v2;
-			}
-			else if (Avx2.IsSupported)
-			{
-				var v1 = Vector256.Create(left._p0, left._p1, left._p2, left._p3).AsByte();
-				var v2 = Vector256.Create(right._p0, right._p1, right._p2, right._p3).AsByte();
-				var equals = Avx2.CompareEqual(v1, v2);
-				var result = Avx2.MoveMask(equals);
-				return (result & 0xFFFF_FFFF) == 0xFFFF_FFFF;
-			}
-			else
-			{
-				return (left._p3 == right._p3) && (left._p2 == right._p2) && (left._p1 == right._p1) && (left._p0 == right._p0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static bool operator !=(in Int256 left, in Int256 right)
-		{
-			if (Vector256.IsHardwareAccelerated)
-			{
-				var v1 = Vector256.Create(left._p0, left._p1, left._p2, left._p3);
-				var v2 = Vector256.Create(right._p0, right._p1, right._p2, right._p3);
-				return v1 != v2;
-			}
-			else if (Avx2.IsSupported)
-			{
-				var v1 = Vector256.Create(left._p0, left._p1, left._p2, left._p3).AsByte();
-				var v2 = Vector256.Create(right._p0, right._p1, right._p2, right._p3).AsByte();
-				var equals = Avx2.CompareEqual(v1, v2);
-				var result = Avx2.MoveMask(equals);
-				return (result & 0xFFFF_FFFF) != 0xFFFF_FFFF;
-			}
-			else
-			{
-				return (left._p3 != right._p3) || (left._p2 != right._p2) || (left._p1 != right._p1) || (left._p0 != right._p0);
-			}
-		}
-
-		/// <inheritdoc/>
-		public static bool operator <(in Int256 left, in Int256 right)
-		{
-			// Successively compare each part.
-			// If left and right have different signs: Signed comparison of _p3 gives result since it is stored as two's complement
-			// If signs are equal and left._p3 < right._p3: left < right for negative and positive values,
-			//                                                    since _p3 is upper 64 bits in two's complement.
-			// If signs are equal and left._p3 > right._p3: left > right for negative and positive values,
-			//                                                    since _p3 is upper 64 bits in two's complement.
-			// If left._p3 == right._p3: unsigned comparison of lower bits gives the result for both negative and positive values since
-			//                                 lower values are lower 64 bits in two's complement.
-			return ((long)left._p3 < (long)right._p3)
-				|| (left._p3 == right._p3 && ((left._p2 < right._p2)
-				|| (left._p2 == right._p2 && ((left._p1 < right._p1)
-				|| (left._p1 == right._p1 && (left._p0 < right._p0))))));
-		}
-
-		/// <inheritdoc/>
-		public static bool operator >(in Int256 left, in Int256 right)
-		{
-			return ((long)left._p3 > (long)right._p3)
-				|| (left._p3 == right._p3 && ((left._p2 > right._p2)
-				|| (left._p2 == right._p2 && ((left._p1 > right._p1)
-				|| (left._p1 == right._p1 && (left._p0 > right._p0))))));
-		}
-
-		/// <inheritdoc/>
-		public static bool operator <=(in Int256 left, in Int256 right)
-		{
-			return ((long)left._p3 < (long)right._p3)
-				|| (left._p3 == right._p3 && ((left._p2 < right._p2)
-				|| (left._p2 == right._p2 && ((left._p1 < right._p1)
-				|| (left._p1 == right._p1 && (left._p0 <= right._p0))))));
-		}
-
-		/// <inheritdoc/>
-		public static bool operator >=(in Int256 left, in Int256 right)
-		{
-			return ((long)left._p3 > (long)right._p3)
-				|| (left._p3 == right._p3 && ((left._p2 > right._p2)
-				|| (left._p2 == right._p2 && ((left._p1 > right._p1)
-				|| (left._p1 == right._p1 && (left._p0 >= right._p0))))));
-		}
 	}
 }
