+++ src/MissingValues/Int256.cs
@@ -4,12 +4,10 @@ using System.Buffers;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
-using System.Numerics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.Intrinsics;
 using System.Text.Json.Serialization;
-using MissingValues.Primitives;
 
 namespace MissingValues
 {
@@ -23,6 +21,12 @@ namespace MissingValues
 	public readonly partial struct Int256
 	{
 		internal const int Size = 32;
+		
+		private static UInt128 _upperMin => new UInt128(0x8000_0000_0000_0000, 0x0000_0000_0000_0000);
+		private static UInt128 _lowerMin => new UInt128(0x0000_0000_0000_0000, 0x0000_0000_0000_0000);
+
+		private static UInt128 _upperMax => new UInt128(0x7FFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF);
+		private static UInt128 _lowerMax => new UInt128(0xFFFF_FFFF_FFFF_FFFF, 0xFFFF_FFFF_FFFF_FFFF);
 
 		/// <summary>
 		/// Represents the value <c>1</c> of the type.
@@ -283,716 +287,6 @@ namespace MissingValues
 		{
 			return TryParse(s, CultureInfo.CurrentCulture, out result);
 		}
-
-		#region From Int256
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="char"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator char(in Int256 value) => (char)value._p0;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="char"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="char"/>.</exception>
-		public static explicit operator checked char(in Int256 value)
-		{
-			if (value._p3 != 0 || value._p2 != 0 || value._p1 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return checked((char)value._p0);
-		}
-		// Signed
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="sbyte"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator sbyte(in Int256 value) => (sbyte)value._p0;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="sbyte"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="sbyte"/>.</exception>
-		public static explicit operator checked sbyte(in Int256 value)
-		{
-			if (~(value._p3 | value._p2 | value._p1) == 0)
-			{
-				long lower = (long)value._p0;
-				return checked((sbyte)lower);
-			}
-
-			if (value._p3 != 0 || value._p2 != 0 || value._p1 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return checked((sbyte)value._p0);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="short"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator short(in Int256 value) => (short)value._p0;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="short"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="short"/>.</exception>
-		public static explicit operator checked short(in Int256 value)
-		{
-			if (~(value._p3 | value._p2 | value._p1) == 0)
-			{
-				long lower = (long)value._p0;
-				return checked((short)lower);
-			}
-
-			if (value._p3 != 0 || value._p2 != 0 || value._p1 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return checked((short)value._p0);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="int"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator int(in Int256 value) => (int)value._p0;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="int"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="int"/>.</exception>
-		public static explicit operator checked int(in Int256 value)
-		{
-			if (~(value._p3 | value._p2 | value._p1) == 0)
-			{
-				long lower = (long)value._p0;
-				return checked((int)lower);
-			}
-
-			if (value._p3 != 0 || value._p2 != 0 || value._p1 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return checked((int)value._p0);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="long"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator long(in Int256 value) => (long)value._p0;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="long"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="long"/>.</exception>
-		public static explicit operator checked long(in Int256 value)
-		{
-			if (~(value._p3 | value._p2 | value._p1) == 0)
-			{
-				return (long)value._p0;
-			}
-
-			if (value._p3 != 0 || value._p2 != 0 || value._p1 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return checked((long)value._p0);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="Int128"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int128(in Int256 value) => (Int128)value.Lower;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="Int128"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="Int128"/>.</exception>
-		public static explicit operator checked Int128(in Int256 value)
-		{
-			if (~(value._p3 | value._p2) == 0)
-			{
-				return new Int128(value._p1, value._p0);
-			}
-
-			if (value._p3 != 0 || value._p2 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return checked((Int128)value.Lower);
-		}
-		/// <summary>
-		/// Implicitly converts a <see cref="Int256" /> value to a <see cref="Int512"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static implicit operator Int512(in Int256 value)
-		{
-			long v = unchecked((long)value._p3);
-			ulong lowerShifted = unchecked((ulong)(v >> 63));
-			return new Int512(
-				lowerShifted, lowerShifted, lowerShifted, lowerShifted, 
-				value._p3, value._p2, value._p1, value._p0
-				);
-		}
-
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="nint"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator nint(in Int256 value) => (nint)value._p0;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="nint"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="nint"/>.</exception>
-		public static explicit operator checked nint(in Int256 value)
-		{
-			if (~(value._p3 | value._p2 | value._p1) == 0)
-			{
-				long lower = (long)value._p0;
-				return checked((nint)lower);
-			}
-
-			if (value._p3 != 0 || value._p2 != 0 || value._p1 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return checked((nint)value._p0);
-		}
-		// Unsigned
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="byte"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator byte(in Int256 value) => (byte)value._p0;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="byte"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="byte"/>.</exception>
-		public static explicit operator checked byte(in Int256 value)
-		{
-			if (value._p3 != 0 || value._p2 != 0 || value._p1 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return checked((byte)value._p0);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="ushort"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator ushort(in Int256 value) => (ushort)value._p0;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="ushort"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="ushort"/>.</exception>
-		public static explicit operator checked ushort(in Int256 value)
-		{
-			if (value._p3 != 0 || value._p2 != 0 || value._p1 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return checked((ushort)value._p0);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="uint"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator uint(in Int256 value) => (uint)value._p0;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="uint"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="uint"/>.</exception>
-		public static explicit operator checked uint(in Int256 value)
-		{
-			if (value._p3 != 0 || value._p2 != 0 || value._p1 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return checked((uint)value._p0);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="ulong"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator ulong(in Int256 value) => value._p0;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="ulong"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="ulong"/>.</exception>
-		public static explicit operator checked ulong(in Int256 value)
-		{
-			if (value._p3 != 0 || value._p2 != 0 || value._p1 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return value._p0;
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="UInt128"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator UInt128(in Int256 value) => value.Lower;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="UInt128"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="UInt128"/>.</exception>
-		public static explicit operator checked UInt128(in Int256 value)
-		{
-			if (value._p3 != 0 || value._p2 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return value.Lower;
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="UInt256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator UInt256(in Int256 value) => Unsafe.BitCast<Int256, UInt256>(value);
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="UInt256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="UInt256"/>.</exception>
-		public static explicit operator checked UInt256(in Int256 value)
-		{
-			if ((long)value._p3 < 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return Unsafe.BitCast<Int256, UInt256>(value);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="UInt512"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator UInt512(in Int256 value) => new(0, 0, 0, 0, value._p3, value._p2, value._p1, value._p0);
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="UInt512"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="UInt512"/>.</exception>
-		public static explicit operator checked UInt512(in Int256 value)
-		{
-			if ((long)value._p3 < 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return new(0, 0, 0, 0, value._p3, value._p2, value._p1, value._p0);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="nuint"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator nuint(in Int256 value) => (nuint)value._p0;
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="nuint"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="nuint"/>.</exception>
-		public static explicit operator checked nuint(in Int256 value)
-		{
-			if (value._p3 != 0 || value._p2 != 0 || value._p1 != 0)
-			{
-				Thrower.IntegerOverflow();
-			}
-			return checked((nuint)value._p0);
-		}
-
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="BigInteger"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator BigInteger(in Int256 value)
-		{
-			if (~(value._p3 & value._p2 & value._p1) == 0)
-			{
-				return new BigInteger((long)value._p0);
-			}
-			if (value._p3 == 0 && value._p2 == 0 && value._p1 == 0)
-			{
-				return new BigInteger(value._p0);
-			}
-
-			Span<byte> span = stackalloc byte[Size];
-			BinaryOperations.WriteInt256LittleEndian(span, in value);
-			return new BigInteger(span, (long)value._p3 >= 0);
-		}
-		// Floating
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="decimal"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator decimal(in Int256 value)
-		{
-			if ((long)value._p3 < 0)
-			{
-				Int256 abs = -value;
-				return -(decimal)(double)(UInt256)(abs);
-			}
-			return (decimal)(double)(UInt256)(value);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="Octo"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Octo(in Int256 value)
-		{
-			if ((long)value._p3 < 0)
-			{
-				Int256 abs = -value;
-				return -(Octo)(UInt256)(abs);
-			}
-			return (Octo)(UInt256)(value);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="Quad"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Quad(in Int256 value)
-		{
-			if ((long)value._p3 < 0)
-			{
-				Int256 abs = -value;
-				return -(Quad)(UInt256)(abs);
-			}
-			return (Quad)(UInt256)(value);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="double"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator double(in Int256 value)
-		{
-			if ((long)value._p3 < 0)
-			{
-				Int256 abs = -value;
-				return -(double)(UInt256)(abs);
-			}
-			return (double)(UInt256)(value);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="float"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator float(in Int256 value)
-		{
-			if ((long)value._p3 < 0)
-			{
-				Int256 abs = -value;
-				return -(float)(UInt256)(abs);
-			}
-			return (float)(UInt256)(value);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="Int256" /> value to a <see cref="Half"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Half(in Int256 value)
-		{
-			if ((long)value._p3 < 0)
-			{
-				Int256 abs = -value;
-				return -(Half)(UInt256)(abs);
-			}
-			return (Half)(UInt256)(value);
-		}
-		#endregion
-		#region To Int256
-		// Signed
-		/// <summary>
-		/// Implicitly converts a <see cref="sbyte" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static implicit operator Int256(sbyte value)
-		{
-			long lower = value;
-			long lowerShifted = lower >> 63;
-			return new((ulong)(lowerShifted), (ulong)lowerShifted, (ulong)lowerShifted, (ulong)lower);
-		}
-		/// <summary>
-		/// Implicitly converts a <see cref="short" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static implicit operator Int256(short value)
-		{
-			long lower = value;
-			long lowerShifted = lower >> 63;
-			return new((ulong)(lowerShifted), (ulong)lowerShifted, (ulong)lowerShifted, (ulong)lower);
-		}
-		/// <summary>
-		/// Implicitly converts a <see cref="int" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static implicit operator Int256(int value)
-		{
-			long lower = value;
-			long lowerShifted = lower >> 63;
-			return new((ulong)(lowerShifted), (ulong)lowerShifted, (ulong)lowerShifted, (ulong)lower);
-		}
-		/// <summary>
-		/// Implicitly converts a <see cref="nint" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static implicit operator Int256(nint value)
-		{
-			long lower = value;
-			long lowerShifted = lower >> 63;
-			return new((ulong)(lowerShifted), (ulong)lowerShifted, (ulong)lowerShifted, (ulong)lower);
-		}
-		/// <summary>
-		/// Implicitly converts a <see cref="long" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static implicit operator Int256(long value)
-		{
-			long lower = value;
-			long lowerShifted = lower >> 63;
-			return new((ulong)(lowerShifted), (ulong)lowerShifted, (ulong)lowerShifted, (ulong)lower);
-		}
-		/// <summary>
-		/// Implicitly converts a <see cref="Int128" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static implicit operator Int256(Int128 value)
-		{
-			ulong lowerShifted = (ulong)((long)value.Upper >> 63);
-			return new(lowerShifted, lowerShifted, value.Upper, value.Lower);
-		}
-		//Unsigned
-		/// <summary>
-		/// Explicitly converts a <see cref="byte" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int256(byte value) => new Int256(0, 0, 0, value);
-		/// <summary>
-		/// Explicitly converts a <see cref="ushort" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int256(ushort value) => new Int256(0, 0, 0, value);
-		/// <summary>
-		/// Explicitly converts a <see cref="uint" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int256(uint value) => new Int256(0, 0, 0, value);
-		/// <summary>
-		/// Explicitly converts a <see cref="nuint" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int256(nuint value) => new Int256(0, 0, 0, value);
-		/// <summary>
-		/// Explicitly converts a <see cref="ulong" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int256(ulong value) => new Int256(0, 0, 0, value);
-		/// <summary>
-		/// Explicitly converts a <see cref="UInt128" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int256(UInt128 value)
-		{
-			return new Int256(0, 0, value.Upper, value.Lower);
-		}
-
-		/// <summary>
-		/// Explicitly converts a <see cref="BigInteger" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int256(BigInteger value)
-		{
-			bool isUnsigned = BigInteger.IsPositive(value);
-
-			Span<byte> span = stackalloc byte[value.GetByteCount()];
-			value.TryWriteBytes(span, out int bytesWritten, isUnsigned);
-
-			if (bytesWritten >= Size)
-			{
-				return BinaryOperations.ReadInt256LittleEndian(span);
-			}
-
-			Int256 result = Zero;
-
-			for (int i = 0; i < bytesWritten; i++)
-			{
-				Int256 part = span[i];
-				part <<= (i * 8);
-				result |= part;
-			}
-
-			result <<= ((Size - bytesWritten) * 8);
-
-			if (!isUnsigned)
-			{
-				result |= ((One << ((Size * 8) - 1)) >> (((Size - bytesWritten) * 8) - 1));
-			}
-
-			return result;
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="BigInteger" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="Int256"/>.</exception>
-		public static explicit operator checked Int256(BigInteger value)
-		{
-			bool isUnsigned = BigInteger.IsPositive(value);
-
-			Span<byte> span = stackalloc byte[isUnsigned ? Size : value.GetByteCount()];
-			if (!value.TryWriteBytes(span, out int bytesWritten, isUnsigned))
-			{
-				Thrower.IntegerOverflow();
-			}
-
-			// Propagate the most significant bit so we have `0` or `-1`
-			sbyte sign = (sbyte)(span[^1]);
-			sign >>= 31;
-
-			// We need to also track if the input data is unsigned
-			isUnsigned |= (sign == 0);
-
-			if (isUnsigned && sbyte.IsNegative(sign) && (bytesWritten >= Size))
-			{
-				// When we are unsigned and the most significant bit is set, we are a large positive
-				// and therefore definitely out of range
-
-				Thrower.IntegerOverflow();
-			}
-
-			if (bytesWritten > Size)
-			{
-				if (span[Size..].IndexOfAnyExcept((byte)sign) >= 0)
-				{
-					// When we are unsigned and have any non-zero leading data or signed with any non-set leading
-					// data, we are a large positive/negative, respectively, and therefore definitely out of range
-
-					Thrower.IntegerOverflow();
-				}
-
-				if (isUnsigned == sbyte.IsNegative((sbyte)span[Size - 1]))
-				{
-					// When the most significant bit of the value being set/clear matches whether we are unsigned
-					// or signed then we are a large positive/negative and therefore definitely out of range
-
-					Thrower.IntegerOverflow();
-				}
-			}
-
-			if (bytesWritten >= Size)
-			{
-				return BinaryOperations.ReadInt256LittleEndian(span);
-			}
-
-			Int256 result = Zero;
-
-			// We have between 1 and 63 bytes, so construct the relevant value directly
-			// since the data is in Little Endian format, we can just read the bytes and
-			// shift left by 8-bits for each subsequent part, then reverse endianness to
-			// ensure the order is correct. This is more efficient than iterating in reverse
-			// due to current JIT limitations
-
-			for (int i = 0; i < bytesWritten; i++)
-			{
-				Int256 part = span[i];
-				part <<= (i * 8);
-				result |= part;
-			}
-
-			result <<= ((Size - bytesWritten) * 8);
-			result = BinaryOperations.ReverseEndianness(in result);
-
-			if (!isUnsigned)
-			{
-				result |= ((One << ((Size * 8) - 1)) >> (((Size - bytesWritten) * 8) - 1));
-			}
-
-			return result;
-		}
-
-		//Floating
-		/// <summary>
-		/// Explicitly converts a <see cref="decimal" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int256(decimal value) => (Int256)(double)value;
-		/// <summary>
-		/// Explicitly converts a <see cref="decimal" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="Int256"/>.</exception>
-		public static explicit operator checked Int256(decimal value) => checked((Int256)(double)value);
-		/// <summary>
-		/// Explicitly converts a <see cref="double" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int256(double value)
-		{
-			const double TwoPow255 = 57896044618658097711785492504343953926634992332820282019728792003956564819968.0;
-
-			if (value <= -TwoPow255)
-			{
-				return MinValue;
-			}
-			else if (double.IsNaN(value))
-			{
-				return 0;
-			}
-			else if (value >= +TwoPow255)
-			{
-				return MaxValue;
-			}
-
-			return ToInt256(value);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="double" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="Int256"/>.</exception>
-		public static explicit operator checked Int256(double value)
-		{
-			const double TwoPow255 = 57896044618658097711785492504343953926634992332820282019728792003956564819968.0;
-
-			if ((0.0d > value + TwoPow255) || double.IsNaN(value) || (value > +TwoPow255))
-			{
-				Thrower.IntegerOverflow();
-			}
-			if (0.0 == TwoPow255 - value)
-			{
-				return MaxValue;
-			}
-
-			return ToInt256(value);
-		}
-		/// <summary>
-		/// Explicitly converts a <see cref="float" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int256(float value) => (Int256)(double)value;
-		/// <summary>
-		/// Explicitly converts a <see cref="float" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="Int256"/>.</exception>
-		public static explicit operator checked Int256(float value) => checked((Int256)(double)value);
-		/// <summary>
-		/// Explicitly converts a <see cref="Half" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		public static explicit operator Int256(Half value) => (Int256)(double)value;
-		/// <summary>
-		/// Explicitly converts a <see cref="Half" /> value to a <see cref="Int256"/>.
-		/// </summary>
-		/// <param name="value">The value to convert.</param>
-		/// <exception cref="OverflowException"><paramref name="value"/> is outside the range of <see cref="Int256"/>.</exception>
-		public static explicit operator checked Int256(Half value) => checked((Int256)(double)value);
-		#endregion
 
 		private static Int256 ToInt256(double value)
 		{
